+++ R/01_fpstruct.R
@@ -334,7 +334,7 @@ is_paragraph <- function(x) {
   chunk_str_names <- c("txt", "font.size", "italic", "bold", "underlined", "strike", "color",
                        "shading.color", "font.family", "hansi.family", "eastasia.family",
                        "cs.family", "vertical.align", "width", "height", "url", "eq_data",
-                       "word_field_data", "img_data",
+                       "word_field_data", "qmd_data", "img_data",
                        ".chunk_index")
   is.data.frame(x) &&
     all(colnames(x) %in% chunk_str_names)
@@ -359,7 +359,7 @@ append_chunkset_struct_element <- function(x, i, j, chunk_data, last = TRUE) {
   chunk_str_names <- c("txt", "font.size", "italic", "bold", "underlined", "strike", "color",
                    "shading.color", "font.family", "hansi.family", "eastasia.family",
                    "cs.family", "vertical.align", "width", "height", "url", "eq_data",
-                   "word_field_data", "img_data")
+                   "word_field_data", "qmd_data", "img_data")
   stopifnot(
     is.data.frame(chunk_data),
     all(chunk_str_names %in% colnames(chunk_data))